requirements.txt
aiogram
Flask
python-dotenv